requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
Werkzeug==3.0.3
gunicorn==22.0.0
python-dotenv==1.0.1
Markdown==3.6
Pillow==10.4.0
